requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20
pyodbc==4.0.39
azure-identity==1.14.0
azure-appconfiguration==1.2.0
python-dotenv==1.0.0
